Move DevUI Config resolution to Runtime
