+++ extensions/vertx-http/runtime/src/main/java/io/quarkus/devui/runtime/config/ConfigDescriptionBean.java
@@ -1,0 +1,17 @@
+package io.quarkus.devui.runtime.config;
+
+import java.util.List;
+
+import io.quarkus.vertx.http.runtime.devmode.ConfigDescription;
+
+public class ConfigDescriptionBean {
+    private List<ConfigDescription> allConfig;
+
+    public ConfigDescriptionBean(final List<ConfigDescription> allConfig) {
+        this.allConfig = allConfig;
+    }
+
+    public List<ConfigDescription> getAllConfig() {
+        return allConfig;
+    }
+}
